automapPlot {automap}	R Documentation

Special plot function for automap

Description

This function wraps around spplot and creates a blue-to-whitish colorscale instead of the standard bpy colorscale.

Usage

automapPlot(plot_data, 
	    zcol,  
	    col.regions, 
	    ...)

Arguments

plot_data
A spatial object that is to be plotted

zcol
The name of the column from plot_data you want to use. Can also be a list.

col.regions
Choose a colors that specify the fill colours.

...
arguments that are passed on to spplot. A sp.layout object for example.

Details

A good function to calculate the position of the colorbreaks the classIntervals function from the classInt package.

Author(s)

Paul Hiemstra, paul@numbertheory.nl

See Also

spplot, plot.autoKrige, plot.posPredictionInterval

Examples

# Ordinary kriging
data(meuse)
coordinates(meuse) =~ x+y
data(meuse.grid)
gridded(meuse.grid) =~ x+y

kriging_result = autoKrige(zinc~1, meuse, meuse.grid)

# Adding the sp.layout parameter shows the locations of the measurements
automapPlot(kriging_result$krige_output, "var1.pred", 
	sp.layout = list("sp.points", meuse))

[Package automap version 1.0-14 Index]