plot.autoKrige {automap}	R Documentation

Plot methods in automap

Description

Defines methods to plot objects in automap.

Usage

## S3 method for class 'autoKrige'
plot(x, sp.layout = NULL, ...)
## S3 method for class 'posPredictionInterval'
plot(x, sp.layout = NULL, justPosition = TRUE, main = "Position prediction interval", ...)

Arguments

x
the object to plot (of class autoKrige or posPredictionInterval)

sp.layout
An object that can contain lines, points and polygons that function as extra layout.

justPosition
logical, if FALSE: not only the plot with the position of the prediction interval is plotted, but also plots with the upper and lower limits of the prediction interval.

main
Title of the plot for the position of the prediction interval.

...
arguments passed to lattice functions xyplot and spplot

Details

For a detailed description of how sp.layout is constructed see spplot.

Author(s)

Paul Hiemstra, paul@numbertheory.nl

See Also

spplot, autoKrige, posPredictionInterval

Examples

# Ordinary kriging
data(meuse)
coordinates(meuse) =~ x+y
data(meuse.grid)
gridded(meuse.grid) =~ x+y

kriging_result = autoKrige(log(zinc)~1, meuse, meuse.grid)
# Adding the sp.layout parameter shows the locations of the measurements
plot(kriging_result, sp.layout = list(pts = list("sp.points", meuse)))

[Package automap version 1.0-14 Index]